feat(experimental): typescript typings
