+++ modules/experimental/package.json
@@ -16,6 +16,7 @@
     "debug",
     "3d"
   ],
+  "types": "src/index.d.ts",
   "main": "dist/es5/index.js",
   "module": "dist/esm/index.js",
   "esnext": "dist/es6/index.js",
