+++ modules/experimental/src/gltf/create-gltf-model.d.ts
@@ -1,0 +1,3 @@
+import ModelNode from "../scenegraph/model-node";
+
+export default function createGLTFModel(gl: WebGLRenderingContext, options: any): ModelNode;
