+++ modules/experimental/src/gltf/create-gltf-objects.d.ts
@@ -1,0 +1,10 @@
+import GLTFAnimator from "./gltf-animator";
+
+export default function createGLTFObjects(
+  gl: WebGLRenderingContext,
+  gltf: any,
+  options?: any
+): {
+  scenes: any;
+  animator: GLTFAnimator;
+};
