+++ modules/experimental/src/gpgpu/histopyramid/histopyramid-shaders.d.ts
@@ -1,0 +1,5 @@
+export const HISTOPYRAMID_BUILD_VS_UTILS: string;
+export const HISTOPYRAMID_BUILD_VS: string;
+export const HISTOPYRAMID_BASE_BUILD_VS: string;
+export const HISTOPYRAMID_TRAVERSAL_UTILS: string;
+export const HISTOPYRAMID_TRAVERSAL_VS: string;
