+++ modules/experimental/src/gpgpu/point-in-polygon/gpu-point-in-polygon.md
@@ -21,7 +21,7 @@ const points = [
 ];
 
 const positionBuffer = new Buffer(gl2, new Float32Array(points));
-const count = points.length / 2;
+const count = 6;
 // Allocate result buffer with enough space (2 floats for each point)
 const filterValueIndexBuffer = new Buffer(gl2, count * 2 * 4);
