+++ modules/experimental/src/gpgpu/point-in-polygon/texture-filter.d.ts
@@ -1,0 +1,8 @@
+
+declare const _default: {
+  name: string;
+  vs: string;
+  getUniforms: (opts) => {};
+};
+
+export default _default;
