+++ modules/experimental/src/index.js
@@ -7,9 +7,9 @@ export {default as GLTFEnvironment} from './gltf/gltf-environment';
 export {default as createGLTFObjects} from './gltf/create-gltf-objects';
 
 // Core nodes
-export {default as ScenegraphNode} from './scenegraph/nodes/scenegraph-node';
-export {default as GroupNode} from './scenegraph/nodes/group-node';
-export {default as ModelNode} from './scenegraph/nodes/model-node';
+export {default as ScenegraphNode} from './scenegraph/scenegraph-node';
+export {default as GroupNode} from './scenegraph/group-node';
+export {default as ModelNode} from './scenegraph/model-node';
 
 // GPGPU utilities for luma.gl
 export {
