--- modules/experimental/src/scenegraph/index.js
@@ -1,5 +1,0 @@
-// SCENEGRAPH
-
-export {default as ScenegraphNode} from './nodes/scenegraph-node';
-export {default as GroupNode} from './nodes/group-node';
-export {default as ModelNode} from './nodes/model-node';
