+++ modules/experimental/src/webvr/vr-button.d.ts
@@ -1,0 +1,4 @@
+export function createEnterVRButton(options: {
+  canvas: any;
+  title: any;
+}): HTMLButtonElement;
