+++ tsconfig.json
@@ -42,10 +42,9 @@
     "modules/debug/test/**/*",
     "modules/engine/src/**/*",
     "modules/engine/test/**/*",
-    "modules/experimental/src/**/*",
-    "modules/experimental/test/**/*",
-    "modules/gltools/src/**/*",
-    "modules/gltools/test/**/*",
+    "modules/experimental/src/gltf/gltf-environment.js",
+    "modules/experimental/src/webvr/vr-display.js",
+    "modules/experimental/src/scenegraph/model-node.js",
     "modules/test-utils/src/**/*",
     "modules/test-utils/test/**/*",
     "modules/webgl/src/classes/**/*",
